+++ lilith/src/main/java/de/huxhorn/lilith/services/details/GroovyEventWrapperHtmlFormatter.java
@@ -1,6 +1,6 @@
 /*
  * Lilith - a log event viewer.
- * Copyright (C) 2014 Joern Huxhorn
+ * Copyright (C) 2014-2015 Joern Huxhorn
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,6 +17,7 @@
  */
 package de.huxhorn.lilith.services.details;
 
+import de.huxhorn.lilith.DateTimeFormatters;
 import de.huxhorn.lilith.data.access.AccessEvent;
 import de.huxhorn.lilith.data.eventsource.EventWrapper;
 import de.huxhorn.lilith.data.logging.LoggingEvent;
@@ -30,6 +31,7 @@ import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.io.Serializable;
+import java.net.URL;
 
 public class GroovyEventWrapperHtmlFormatter
 	extends AbstractHtmlFormatter
@@ -90,6 +92,8 @@ public class GroovyEventWrapperHtmlFormatter
 					accessEvent = (AccessEvent) event;
 				}
 
+				URL messageViewRootUrl = applicationPreferences.getDetailsViewRootUrl();
+
 				Binding binding = new Binding();
 
 				binding.setVariable(LOGGER_VARIABLE, logger);
@@ -101,6 +105,10 @@ public class GroovyEventWrapperHtmlFormatter
 				binding.setVariable(COMPLETE_CALL_STACK_OPTION_VARIABLE, applicationPreferences.isShowingFullCallstack());
 				binding.setVariable(SHOW_STACK_TRACE_OPTION_VARIABLE, applicationPreferences.isShowingStackTrace());
 				binding.setVariable(WRAPPED_EXCEPTION_STYLE_OPTION_VARIABLE, applicationPreferences.isUsingWrappedExceptionStyle());
+
+				binding.setVariable(DOCUMENT_ROOT_VARIABLE, messageViewRootUrl.toExternalForm());
+
+				binding.setVariable(DATETIME_FORMATTER_VARIABLE, DateTimeFormatters.DATETIME_IN_SYSTEM_ZONE_SPACE);
 
 				instance.setBinding(binding);
 				Object result = instance.run();